Move Watching to its own file
